+++ src/a03_01_arvore_binaria/ExpressionTool.java
@@ -2,15 +2,15 @@ package a03_01_arvore_binaria;
 
 public class ExpressionTool {
 	
-	public static Integer evaluate(BTNode root) {
+	public static Double evaluate(BTNode root) {
 		if (root != null) {
 			// Se a informação já for um número (folhas), retorna-o
-			if (root.getInfo() instanceof Integer) {
-				return (Integer) root.getInfo();
+			if (root.getInfo() instanceof Double) {
+				return (Double) root.getInfo();
 			}
 			// Avalia os nós
-			Integer left = evaluate(root.getLeft());
-			Integer right = evaluate(root.getRight());
+			Double left = evaluate(root.getLeft());
+			Double right = evaluate(root.getRight());
 			// Executar a operação com o resultado dos nós
 			if (root.getInfo().equals("+")) {
 				return left + right;
